Output File I/O error message via GUI instead of System.out
The error message when updating task list changes to save disk is
displayed via System.out instead of the GUI.

As we transition to a GUI app, user is unlikely to monitor the
standard output. Thus, user may miss the error message.

Let's change the File I/O error message to be shown as a bot
message in the JavaFX GUI.